:gear: Extraire les mécanismes dans un fichier dédié
